refactor(pipeline): decompose the observation pipeline god-file by role (WS-D)
Split internal/alerts/pipeline/observation_pipeline.go (669 LOC) into
role-based files in the same package — pure move: pipeline lifecycle stays
in observation_pipeline.go; history priming moves to observation_prime.go;
per-kind scan/evaluation moves to observation_eval.go.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
